requirements.txt
python-telegram-bot[webhooks]==21.10
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
Pillow==11.1.0
python-dotenv==1.0.1
